+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes qsf-integration project will be documented in this file.
 - add userId and sessionId to the request and tracking
 - add a method to get a ObjectNode in the JsonBuilder
 - support date_histogram facets with a custom query
+- add a field name to the search filter
 
 ### Changed
 - use the document id as a fallback for tracking returned documents